Major update + but fix
Add Debug mode
Add LED and Keypad Factory QC check mode
Modified LED xy HAL API

+++ src/parameter/MatrixVariable.cpp
@@ -5,7 +5,7 @@ u8 device_id = 0; //0~255 0 for off
 
 //LED setting
 u8 brightness = 32;  //VeryDim - 32 | Dim - 64 | Normal(Default) - 96 | Bright - 128 | VeryBright - 160 | Bright AF - 192  (Higher then 192 could result in USB resettable fuse disconnect(500ma))
-u64 palette[3][128] =     //WRGB Colour Palette
+u32 palette[3][128] =     //WRGB Colour Palette
 {{            //MatrixColorPalette
   0x00000000, //0
   0x00050505, //1
